CPU protection: works, make it work better
Had a glitch (to be fixed in a later commit) which the CPU protection
against busy-spinning loops on failure protected against system
thrashing, but it still logged every second.  We can do better.

Exponential backoff for spinning scenarios, with reset upon success.
Log the sleeps.
Avoid sleeping on first iteration, because with logging it's no longer
acceptable.