Timer expiration fixes: scan files, set earliest right
Timer expiration for a _directory_ led to failures on every scan,
because we were passing a list of file paths to a function for scanning
command-line paths and which obeyed the `r.config.Directories` bool.

Switch to a method which takes a callback function and pass it a bound
method in both places, one of which was the old
dispatch-on-one-path-from-commandline method and the timer-based
invocation using the "must be a file, is real error if not OCSP" method.

Spotted two bugs in figuring out the next dispatch times:
 * If we swept to the end of all certs (normal for singular list) we
   wouldn't figure out the new start time;
 * A newly registered time can be before any times from existing certs,
   if we have one cert with short timers and one cert with long timers,
   so the short-timer-renewed case is a candidate for being first.

Solve both by being less clever: always look at all current timers to
figure out the earliest.