Don't "keep" an expired scheduled time
When we hit scheduled check at time Ts, and schedule a future check at
time Tf, because we already have a sooner check, we ignored the
extension.  Because we didn't _remove_ the time when scheduling it.

Don't keep a scheduled time in the past.  The CPU-protecting exponential
backoff is what saved us.