Ease semantic analysis: don't return item in err case
When the HTTP result is not OK, we return `ErrHTTPFailure` but also return the
potentially-large payload.  But the caller, upon seeing the error, does nothing
with the payload before erroring out and then losing the object.

So no bug.

But at least one semantic analysis tool missed that the normal end-of-function
return would not be reached in this case, and gave a false positive.

In this case, removing the option to inspect and log the non-200 response is
not a big loss, so make life easier for tooling.