FreeBSD: change daemon startup for better logging
These days daemon(8) has a `-o` option to write to a logfile; the default mode
redirects both stdout and stderr, which is what I want.  In addition, there is
extensive buffering of daemon's own stdout/stderr whereas the output from the
child seems to make it through to the -o file with line-buffering intact.

So use the -o and get log output in a timely manner.